Add WIP of heavy refactoring second step
1. This is Work in progress (WIP). Do NOT MERGE THIS INTO MASTER.

2. Refactor the paho MQTT specific stuff into PahoAdapter,
   which can then be used for interacting with the broker and also
   mocking/stubbing it in actual tests related to this project.

   Having to setup 1:1 a full real life installation with remote
   clients, a messaging queue (paho broker) and so on isn't the
   brightest idea to say the least when writing tests for a project.

   So I did the first important task of separating stuff which shouldn't
   be too intermingled. May it be for testing purposes, which is why
   I do it and also for understandability of the project.

3. Interaction of player with the paho mqtt client is now indirectly
   done by find_client_for_player method on the paho adapter.
   So I get the same functionality, without intermixing stuff which
   will be then after a couple of months regretted to be intermingled.

4. Moves code out of ghoust_srv.py and into ghoust package.

5. The next step is to clean up that refactor and actually write the
   unit/integration tests without depdening on a running paho instance.
